big update: added minion class; slightly lowered overall difficulty; many values tweaked
after having watched a few different people test my game, i've found that pplayers quickly become overwhelmed and are not sure what to do even after having it explained to them, so i've decided to lower the overall difficulty in a number of ways:
1. antibodies now last twice as long (increases healing chances)
2. boss no longer heals itself when eating cells
3. cells get slightly more energy when colliding with each other
the next thing i will change is increasing the maximum number of antibodies on screen at once.

also: added Minion class.
the minions, currently, are basically copy-pasted parasite code that exist on the ends of the boss's bezier appendages. their attack function does not work at the moment but i'll fix it up.
also changed bezier tail behaviour, including attack feedback: the tails will flail wildly as they are being attacked, which is a pretty cool effect.

+++ final_project_B/Parasite.pde
@@ -31,7 +31,7 @@ class Parasite {
   // variable used to get stunned parasites to shake
   int stunOffset = 0;
 
-  Parasite(int tempX, int tempY, int tempSize) {
+  Parasite(float tempX, float tempY, int tempSize) {
     x = tempX;
     y = tempY;
     size = tempSize;
@@ -243,6 +243,9 @@ class Parasite {
   void display() {
     if (eatCount <= 64) {
       fill(fill, 96); 
+      
+      
+      
     }
     else if (eatCount > 64) {
       fill (fed, 96);
